src/third_party/WebKit/LayoutTests/animations/animationworklet/animation-worklet-inside-iframe.html

<!DOCTYPE html>
<style>
.redbox {
  width: 100px;
  height: 100px;
  background-color: #ff0000;
}
</style>
<script id="main_worklet" type="text/worklet">
function CreateLoggingAnimator(name, localTime) {
  return class {
    constructor() {
        console.log(name + " constructor called.");
    }
    animate(currentTime, effect) {
      if (this.animated)
        return;
      this.animated = true;

      console.log(name + " animate called.");
      effect.localTime = localTime;
    }
  }
}

registerAnimator("duplicate_animator", CreateLoggingAnimator("mainframe/duplicate_animator", 1000));
</script>

<script src="resources/animation-worklet-tests.js"></script>
<script>
if (window.testRunner) {
  testRunner.waitUntilDone();
  testRunner.dumpAsText();
}

// Wait for iframe to load and start its animations.
window.onmessage = function(msg) {
  waitTwoAnimationFrames(run_test);
};

function run_test() {
  runInAnimationWorklet(
    document.getElementById('main_worklet').textContent
  ).then(_ => {
    // Create an animation for duplicate animator.
    const animation = new WorkletAnimation('duplicate_animator', new KeyframeEffect(document.getElementById("target"), [{ opacity: 0 }], { duration: 1000 }));
    animation.play();

    if (window.testRunner) {
      waitTwoAnimationFrames( _ => {
        testRunner.notifyDone();
      });
    }
  });
}
</script>

<div id="target" class="redbox"></div>
<iframe src="resources/animator-iframe.html"></iframe>
