third_party/WebKit/LayoutTests/animations/animationworklet/animation-worklet-inside-iframe-expected.txt

CONSOLE MESSAGE: line 5: iframe/iframe_animator constructor called.
CONSOLE MESSAGE: line 12: iframe/iframe_animator animate called.
CONSOLE MESSAGE: line 5: mainframe/duplicate_animator constructor called.
CONSOLE MESSAGE: line 12: mainframe/duplicate_animator animate called.
